Rename CSS files to match with components / pages they are bound to
