+++ components/bridge/ChainPicker.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Chain, SupportedChain } from "@/types";
 
-import styles from "@/styles/components/bridge/chainpicker.module.css";
+import styles from "@/styles/components/bridge/ChainPicker.module.css";
 import { CHAINS } from "@/utils/bridge/helpers";
 
 const ChainPicker: React.FC<{
