+++ styles/components/shared/ExchangeIcon.module.css
@@ -1,4 +1,4 @@
-.swapIcon {
+.exchangeIcon {
 	height: 56px;
 	width: 56px;
 	color: rgba(17, 48, 255, 0.3);
@@ -13,7 +13,7 @@
 	align-items: center;
 }
 
-.swapIcon:hover {
+.exchangeIcon:hover {
 	fill: white;
 	background: #1130ff;
 }
@@ -25,7 +25,7 @@
 	width: 60px;
 }
 
-.swapIcon:hover > img {
+.exchangeIcon:hover > img {
 	filter: invert(1) sepia(1) hue-rotate(200deg);
 }
